requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
pillow
PyYAML>=5.3
scipy>=1.4.1
tensorboard>=2.2
torch==1.7.0
torchvision>=0.7.0
tqdm>=4.41.0
wandb
tabulate
onnx-simplifier==0.2.19
p_tqdm

# coco ----------------------------------------
# pycocotools>=2.0

# export --------------------------------------
packaging  # for coremltools
coremltools==4.0
onnx==1.7.0
scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
seaborn  # plotting
youtube-dl

# optuna
optuna
psycopg2-binary
